+++ bin-doom/p_saveg.c
@@ -24,6 +24,8 @@
 static const char
 rcsid[] = "$Id: p_tick.c,v 1.4 1997/02/03 16:47:55 b1 Exp $";
 
+#include <stdint.h>
+
 #include "i_system.h"
 #include "z_zone.h"
 #include "p_local.h"
@@ -37,7 +39,7 @@ byte*		save_p;
 
 // Pads save_p to a 4-byte boundary
 //  so that the load/save works on SGI&Gecko.
-#define PADSAVEP()	save_p += (4 - ((int) save_p & 3)) & 3
+#define PADSAVEP()	save_p += (4 - ((uintptr_t) save_p & 3)) & 3
 
 
 
@@ -64,8 +66,8 @@ void P_ArchivePlayers (void)
 	{
 	    if (dest->psprites[j].state)
 	    {
-		dest->psprites[j].state 
-		    = (state_t *)(dest->psprites[j].state-states);
+		dest->psprites[j].state
+		    = (state_t *)(intptr_t)(dest->psprites[j].state-states);
 	    }
 	}
     }
@@ -100,8 +102,8 @@ void P_UnArchivePlayers (void)
 	{
 	    if (players[i]. psprites[j].state)
 	    {
-		players[i]. psprites[j].state 
-		    = &states[ (int)players[i].psprites[j].state ];
+		players[i]. psprites[j].state
+		    = &states[ (intptr_t)players[i].psprites[j].state ];
 	    }
 	}
     }
@@ -244,10 +246,10 @@ void P_ArchiveThinkers (void)
 	    mobj = (mobj_t *)save_p;
 	    memcpy (mobj, th, sizeof(*mobj));
 	    save_p += sizeof(*mobj);
-	    mobj->state = (state_t *)(mobj->state - states);
-	    
+	    mobj->state = (state_t *)(intptr_t)(mobj->state - states);
+
 	    if (mobj->player)
-		mobj->player = (player_t *)((mobj->player-players) + 1);
+		mobj->player = (player_t *)(intptr_t)((mobj->player-players) + 1);
 	    continue;
 	}
 		
@@ -299,11 +301,11 @@ void P_UnArchiveThinkers (void)
 	    mobj = Z_Malloc (sizeof(*mobj), PU_LEVEL, NULL);
 	    memcpy (mobj, save_p, sizeof(*mobj));
 	    save_p += sizeof(*mobj);
-	    mobj->state = &states[(int)mobj->state];
+	    mobj->state = &states[(intptr_t)mobj->state];
 	    mobj->target = NULL;
 	    if (mobj->player)
 	    {
-		mobj->player = &players[(int)mobj->player-1];
+		mobj->player = &players[(intptr_t)mobj->player-1];
 		mobj->player->mo = mobj;
 	    }
 	    P_SetThingPosition (mobj);
